Qt6 Support

+++ Source/Libs/ZSDraw/Widgets/GraphObjs/ZSDrawGraphObjGeometryModel.cpp
@@ -580,7 +580,7 @@ int CModelGraphObjGeometry::rowCount(const QModelIndex& i_modelIdxParent) const
         /* strMethodInArgs    */ strMthInArgs );
 
     if (mthTracer.areMethodCallsActive(EMethodTraceDetailLevel::ArgsNormal)) {
-        mthTracer.setMethodReturn(m_arLabelSettings.size());
+        mthTracer.setMethodReturn(static_cast<int>(m_arLabelSettings.size()));
     }
     return m_arLabelSettings.size();
 }
